requirements.txt
cerebras-cloud-sdk
requests
beautifulsoup4
langchain
langchain-community
python-dotenv
duckduckgo-search 